+++ NEWS.md
@@ -1,3 +1,11 @@
-# lstcompr 0.1.0
+# listcompr 0.2.0
+
+* variable ranges can also be of character type (fixes https://github.com/patrickroocks/listcompr/issues/1)
+* fixed sorting of free variables (fixes https://github.com/patrickroocks/listcompr/issues/2)
+* auto-detect column names when generating data frames where a symbol is in the base expression
+* fixed missing export of `gen.logical.or`, added example to the vignette
+* allow substitutions in the list comprehension arguments
+
+# listcompr 0.1.0
 
 * First version of the listcompr package